+++ src/app/Scenes/Playground/Playground.tsx
@@ -1,44 +1,5 @@
-import { Flex, Text } from "@artsy/palette-mobile"
-import App0 from "app/Scenes/Playground/L0-Reanimated"
-import App1 from "app/Scenes/Playground/L1-Animated-Components"
-import App2 from "app/Scenes/Playground/L2-Shared-Values"
-import App3 from "app/Scenes/Playground/L3-Moti"
-import App4 from "app/Scenes/Playground/L4-Gestures"
-import App5 from "app/Scenes/Playground/L5-Exercice"
-
-// eslint-disable-next-line prefer-const
-let LEVEL = 5
+import { Animations } from "app/Scenes/Playground/Animation"
 
 export const Playground = () => {
-  const Content = () => {
-    switch (LEVEL) {
-      case 0:
-        return <App0 />
-      case 1:
-        return <App1 />
-      case 2:
-        return <App2 />
-      case 3:
-        return <App3 />
-      case 4:
-        return <App4 />
-      case 5:
-        return <App5 />
-      default:
-        return null
-    }
-  }
-
-  return (
-    <Flex>
-      <Flex backgroundColor="black10" p={2}>
-        <Text variant="sm-display" fontWeight="bold" textAlign="center">
-          Animations with Reanimated 3
-        </Text>
-      </Flex>
-      {/* <Flex flex={1}> */}
-      <Content />
-      {/* </Flex> */}
-    </Flex>
-  )
+  return <Animations />
 }
